pyproject.toml
[project]
name = "gbd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "distro==1.9.0",
    "exceptiongroup==1.3.0",
    "h11==0.16.0",
    "haystack-ai==2.13.2",
    "haystack-experimental==0.9.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "lazy-imports==0.4.0",
    "markupsafe==3.0.2",
    "more-itertools==10.7.0",
    "networkx==3.4.2",
    "numpy==2.2.5",
    "openai==1.78.1",
    "posthog==4.0.1",
    "psycopg2>=2.9.10",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.24.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
]
